+++ src/interface/web/app/share/chat/layout.tsx
@@ -1,6 +1,6 @@
 import type { Metadata } from "next";
 import { Noto_Sans } from "next/font/google";
-import "../../../globals.css";
+import "../../globals.css";
 
 const inter = Noto_Sans({ subsets: ["latin"] });
